+++ migrations/000001_create_realty_table.up.sql
@@ -14,10 +14,10 @@ CREATE TABLE IF NOT EXISTS realty (
     raw_property_type TEXT NOT NULL,
     property_type TEXT NOT NULL,
     updated TIMESTAMP NOT NULL,
-    rent_range INTEGER[] NOT NULL,
-    beds_range INTEGER[] NOT NULL,
-    baths_range INTEGER[] NOT NULL,
-    dimensions_range INTEGER[] NOT NULL
+    rent_range FLOAT[] NOT NULL,
+    beds_range FLOAT[] NOT NULL,
+    baths_range FLOAT[] NOT NULL,
+    dimensions_range FLOAT[] NOT NULL
 );
 
 CREATE INDEX idx_city ON realty (city_name);
